Unify graph adapter factory
