+++ docs/CONFIG_TEMPLATES.md
@@ -60,6 +60,13 @@ below lists all available variables and their default values.
 | --- | --- | --- |
 | `UME_DB_PATH` | `ume_graph.db` | SQLite database used by `PersistentGraph`. |
 | `UME_GRAPH_BACKEND` | `sqlite` | Backend for graph storage (`sqlite`, `postgres`, `redis`, or `neo4j`). |
+
+Valid values for `UME_GRAPH_BACKEND` are:
+
+- `sqlite` – local SQLite database via `PersistentGraph`.
+- `postgres` – PostgreSQL backend using `PostgresGraph`.
+- `redis` – Redis-backed `RedisGraphAdapter`.
+- `neo4j` – Neo4j graph database via `Neo4jGraph`.
 | `UME_SNAPSHOT_PATH` | `ume_snapshot.json` | Path to graph snapshot file. |
 | `UME_SNAPSHOT_DIR` | `.` | Directory that snapshot APIs will accept paths from. |
 | `UME_AUDIT_LOG_PATH` | `audit.log` | Location of the audit log. |
