+++ src/ume/config/__init__.py
@@ -14,7 +14,7 @@ class Settings(BaseSettings):  # type: ignore[misc]
 
     # UME Core
     UME_DB_PATH: str = "ume_graph.db"
-    UME_GRAPH_BACKEND: str = "sqlite"  # sqlite, postgres, or redis
+    UME_GRAPH_BACKEND: str = "sqlite"  # sqlite, postgres, redis, or neo4j
     UME_SNAPSHOT_PATH: str = "ume_snapshot.json"
     UME_SNAPSHOT_DIR: str = "."
     UME_COLD_DB_PATH: str = "ume_cold.db"
