+++ src/ume/resources.py
@@ -3,24 +3,8 @@
 
 from typing import Callable
 
-from .config import settings
 from .graph_adapter import IGraphAdapter
-from .persistent_graph import PersistentGraph
-from .postgres_graph import PostgresGraph
-from .redis_graph_adapter import RedisGraphAdapter
-from .rbac_adapter import RoleBasedGraphAdapter
-from .tracing import TracingGraphAdapter, is_tracing_enabled
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover - for type hints only
-    from .neo4j_graph import Neo4jGraph
-else:  # pragma: no cover - optional dependency
-    try:
-        from .neo4j_graph import Neo4jGraph
-    except Exception:
-        class Neo4jGraph:
-            def __init__(self, *_: object, **__: object) -> None:
-                raise ImportError("neo4j is required for Neo4jGraph")
+from .factories import create_graph_adapter as _create_base_adapter
 from .vector_store import VectorBackend, create_default_store
 
 
@@ -31,28 +15,7 @@ def create_graph_adapter(
 ) -> IGraphAdapter:
     """Instantiate the configured :class:`IGraphAdapter`."""
 
-    backend = settings.UME_GRAPH_BACKEND.lower()
-    if backend == "postgres":
-        base: IGraphAdapter = PostgresGraph(db_path or settings.UME_DB_PATH)
-    elif backend == "redis":
-        base = RedisGraphAdapter(db_path or settings.UME_DB_PATH)
-    elif backend == "neo4j":
-        base = Neo4jGraph(
-            settings.NEO4J_URI,
-            settings.NEO4J_USER,
-            settings.NEO4J_PASSWORD,
-        )
-    else:
-        base = PersistentGraph(db_path or settings.UME_DB_PATH)
-
-    if is_tracing_enabled():
-        base = TracingGraphAdapter(base)
-
-    role = role if role is not None else settings.UME_ROLE
-    if role:
-        base = RoleBasedGraphAdapter(base, role=role)
-
-    return base
+    return _create_base_adapter(db_path, role=role)
 
 
 def _default_graph_factory() -> IGraphAdapter:
